etu_django_frame/django/2.2.2/requirements/requirements.txt
Django==2.2.2
celery==4.3.0
django-cas-ng==3.6.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.0.2
django-filter==2.2.0
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
djangorestframework==3.11.0
django-simpleui==2021.8.17
drf-yasg==1.17.1
Markdown==3.2.1
ipython==7.11.0
kombu==4.6.4
mysqlclient==1.3.14
redis==3.3.11
requests==2.22.0
supervisor==4.0.3
uWSGI==2.0.18
vine==1.3.0